Удалил лишние инклюды

+++ search-server/request_queue.cpp
@@ -2,13 +2,13 @@
 
 std::vector<Document> RequestQueue::AddFindRequest(const std::string& raw_query, DocumentStatus status) {
     const auto result = search_server_.FindTopDocuments(raw_query, status);
-    AddRequest(result.size());
+    AddRequest(static_cast<int>(result.size()));
     return result;
 }
 
 std::vector<Document> RequestQueue::AddFindRequest(const std::string& raw_query) {
     const auto result = search_server_.FindTopDocuments(raw_query);
-    AddRequest(result.size());
+    AddRequest(static_cast<int>(result.size()));
     return result;
 }
